Moves "handlers" into "utils"
