+++ src/utils/request/getPublicUrlFromRequest.ts
@@ -1,5 +1,5 @@
 import { format } from 'url'
-import { MockedRequest } from '../../handlers/requestHandler'
+import { MockedRequest } from '../handlers/requestHandler'
 
 /**
  * Returns a relative URL if the given request URL is relative to the current origin.
